Data.Validation

Description

A data type similar to Data.Either that accumulates failures.

Data type

data Validation err a

An Validation is either a value of the type err or a, similar to Either. However, the Applicative instance for Validation accumulates errors using a Semigroup on err. In contrast, the Applicative for Either returns only the first error.

A consequence of this is that Validation has no Bind or Monad instance. This is because such an instance would violate the law that a Monad's ap must equal the Applicative's <*>

An example of typical usage can be found here.

Constructors

Failure err

Success a

Constructing validations

validate :: Validate v => e -> (a -> Maybe b) -> a -> v e b

validates an a producing an updated optional value, returning e in the empty case.

This can be thought of as having the less general type:

validate :: e -> (a -> Maybe b) -> a -> Validation e b

validationNel :: Either e a -> Validation (NonEmpty e) a

validationNel is liftError specialised to NonEmpty lists, since they are a common semigroup to use.

fromEither :: Either e a -> Validation e a

Converts from Either to Validation.

liftError :: (b -> e) -> Either b a -> Validation e a

liftError is useful for converting an Either to an Validation when the Left of the Either needs to be lifted into a Semigroup.

Functions on validations

validation :: (e -> c) -> (a -> c) -> Validation e a -> c

validation is the catamorphism for Validation.

toEither :: Validation e a -> Either e a

Converts from Validation to Either.

orElse :: Validate v => v e a -> a -> a

v orElse a returns a when v is Failure, and the a in Success a.

This can be thought of as having the less general type:

orElse :: Validation e a -> a -> a

valueOr :: Validate v => (e -> a) -> v e a -> a

Return the a or run the given function over the e.

This can be thought of as having the less general type:

valueOr :: (e -> a) -> Validation e a -> a

ensure :: Validate v => e -> (a -> Maybe b) -> v e a -> v e b

ensure ensures that a validation remains unchanged upon failure, updating a successful validation with an optional value that could fail with e otherwise.

This can be thought of as having the less general type:

ensure :: e -> (a -> Maybe b) -> Validation e a -> Validation e b

codiagonal :: Validation a a -> a

codiagonal gets the value out of either side.

validationed :: Validate v => (v e a -> v e' a') -> Validation e a -> Validation e' a'

Run a function on anything with a Validate instance (usually Either) as if it were a function on Validation

This can be thought of as having the type

(Either e a -> Either e' a') -> Validation e a -> Validation e' a'

bindValidation :: Validation e a -> (a -> Validation e b) -> Validation e b

bindValidation binds through an Validation, which is useful for composing Validations sequentially. Note that despite having a bind function of the correct type, Validation is not a monad. The reason is, this bind does not accumulate errors, so it does not agree with the Applicative instance.

There is nothing wrong with using this function, it just does not make a valid Monad instance.

Prisms

These prisms are useful for writing code which is polymorphic in its choice of Either or Validation. This choice can then be made later by a user, depending on their needs.

An example of this style of usage can be found here

_Failure :: Validate f => Prism (f e1 a) (f e2 a) e1 e2

This prism generalises _Left. It targets the failure case of either Either or Validation.

_Success :: Validate f => Prism (f e a) (f e b) a b

This prism generalises _Right. It targets the success case of either Either or Validation.

Isomorphisms

class Validate f where

The Validate class carries around witnesses that the type f is isomorphic to Validation, and hence isomorphic to Either.

Minimal complete definition

_Validation

Methods

_Validation :: Iso (f e a) (f g b) (Validation e a) (Validation g b)

_Either :: Iso (f e a) (f g b) (Either e a) (Either g b)

revalidate :: (Validate f, Validate g) => Iso (f e1 s) (f e2 t) (g e1 s) (g e2 t)

revalidate converts between any two instances of Validate.